requirements.txt
amqp==5.1.1
argh==0.26.2
asgiref==3.5.2
async-timeout==4.0.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.6
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.4
django-annoying==0.10.6
django-celery-beat==2.3.0
django-celery-results==2.3.1
django-ckeditor==6.4.1
django-js-asset==2.0.0
django-redis==5.2.0
django-timezone-field==5.0
django-widget-tweaks==1.4.12
gunicorn==20.1.0
kombu==5.2.4
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.29
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
redis==4.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
vine==5.0.0
watchdog==2.1.8
wcwidth==0.2.5
wrapt==1.14.1
